+++ tests/test_app.py
@@ -3,8 +3,32 @@ import re
 import pytest
 from rich._emoji_codes import EMOJI
 
-from cruft_helloworld.app import console, hello_world
-from cruft_helloworld.tools.enums import GlobeEmoji
+from cruft_helloworld.app import (
+    PACKAGE_NAME,
+    PACKAGE_VERSION,
+    cli,
+    console,
+    hello_world,
+)
+
+
+def test_app_cli_show_version(cli_runner):
+    result = cli_runner.invoke(cli, ["--version"])
+    assert result.exit_code == 0
+    application_name_expected = PACKAGE_NAME
+    application_version_expected = PACKAGE_VERSION
+    version_message_expected = (
+        f"{application_name_expected}, version {application_version_expected}\n"
+    )
+    assert result.output == version_message_expected
+
+
+def test_app_cli_show_banner(cli_runner):
+    result = cli_runner.invoke(cli, ["--show-banner"])
+    assert result.exit_code == 0
+    version_message = f"{PACKAGE_NAME}, version {PACKAGE_VERSION}"
+    assert len(result.output) > len(version_message)
+    assert result.output.count("\n") > 5
 
 
 def test_app_cli_help(cli_runner):
@@ -17,7 +41,7 @@ def test_app_cli_help(cli_runner):
 
 
 @pytest.mark.parametrize(
-    "globe_emoji_name,globe_emoji_char",
+    "input_globe_emoji_name,expected_globe_emoji_char",
     (
         ("EUROPE_AFRICA", "🌍"),
         ("americas", "🌎"),
@@ -26,15 +50,15 @@ def test_app_cli_help(cli_runner):
     ),
 )
 def test_app_cli_hello_world(
-    cli_runner, globe_emoji_name: GlobeEmoji, globe_emoji_char: str
+    cli_runner, input_globe_emoji_name: str, expected_globe_emoji_char: str
 ):
     # https://github.com/willmcgugan/rich/blob/a3f5609202e9aa45751ce9baa3a72462ed1cc488/tests/test_console.py#L192
     with console.capture() as capture:
         result = cli_runner.invoke(
-            hello_world, ["--globe-emoji", globe_emoji_name] if globe_emoji_name else []
+            hello_world, ["--globe-emoji", input_globe_emoji_name]
         )
     assert result.exit_code == 0
-    assert capture.get() == f"Hello {globe_emoji_char}\n"
+    assert capture.get() == f"Hello {expected_globe_emoji_char}\n"
 
 
 def test_error_app_cli_hello_world(cli_runner):